Jamar/project04/requirements.txt
fastapi
uvicorn
python-multipart
langchain
pydantic
reportlab
PyPDF2
OpenAi
python-dotenvpi
langchain-openai
chromadb